+++ changelog.md
@@ -1,10 +1,22 @@
 # v1.0.4-beta3 (Big rewrite of the musicbot module)
 
+# beta 3
+- Fixed automatic update rate of getEmbed()
+- Prevented some 400 responses by shortening fields in the musicbot embed
+- Both features toghether should brevent getting rate limited by discord
+- Increased timeframe to check if bot was disconnected to prevent accidental queue stop
+- /add_song now shows an arrow in the queue if the song is in the viewable list
+- You can now use commas for multiple songs with /play and /add_song. eg: /play url=[Song1, Song2] (works with urls, playlists and youtube search)
+- When suggesting a wrong track the player now restarts with the new song
+- When using the suggest button if the song skips the suggested one is added to the queue as next but not played until the current one finishes
+- Fixed issue with youtubeparser (see diff)
+- getEnv() has generally better parsing
+- This version seems to fix most of the big issues with the latest version
+
 # beta 2
 - file bot.py now supports the lang.py file (TODO: easy language change in config.py)
 - new config for music player to wait in VC after the queue finishes
 - other fixes
-
 
 # beta 1
 - lang is now a python file for better management